ProcessSet rised to root class from Stage's child.
